+++ src/main/java/spoticks/ticket_reservation/global/config/SecurityConfig.java
@@ -12,7 +12,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import spoticks.ticket_reservation.global.auth.CustomUserDetailsService;
+import spoticks.ticket_reservation.global.auth.service.CustomUserDetailsService;
 import spoticks.ticket_reservation.global.security.JwtAuthenticationFilter;
 import spoticks.ticket_reservation.global.security.JwtEntryPoint;
